% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodos_emmeans_logistf.R
\name{emm_basis.logistf}
\alias{emm_basis.logistf}
\title{Base para medias marginales estimadas de un modelo logistf}
\usage{
\method{emm_basis}{logistf}(object, trms, xlev, grid, ...)
}
\arguments{
\item{object}{Objeto de clase \code{logistf}}

\item{trms}{Términos del modelo}

\item{xlev}{Niveles de los factores}

\item{grid}{Grid de referencia donde evaluar las medias}

\item{...}{Argumentos adicionales}
}
\value{
Lista con la base para calcular emmeans
}
\description{
Base para medias marginales estimadas de un modelo logistf
}
\keyword{internal}
