% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.modelo_germinacion_unificado}
\alias{print.summary.modelo_germinacion_unificado}
\title{Imprimir Resumen de Modelo}
\usage{
\method{print}{summary.modelo_germinacion_unificado}(x, digits = 4, signif.stars = TRUE, ...)
}
\arguments{
\item{x}{Objeto de clase summary.modelo_germinacion_unificado.}

\item{digits}{Número de decimales.}

\item{signif.stars}{Lógico. Mostrar estrellas de significancia.}

\item{...}{Argumentos adicionales.}
}
\description{
Imprime el resumen de un modelo RegenerAR.
}
