% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodos_emmeans_logistf.R
\name{recover_data.logistf}
\alias{recover_data.logistf}
\title{Recuperar datos de un modelo logistf para emmeans}
\usage{
\method{recover_data}{logistf}(object, trms, xlev, grid, ...)
}
\arguments{
\item{object}{Objeto de clase \code{logistf}}

\item{trms}{Términos del modelo}

\item{xlev}{Niveles de los factores}

\item{grid}{Grid de referencia (no usado aquí)}

\item{...}{Argumentos adicionales}
}
\value{
Objeto de datos recuperados compatible con emmeans
}
\description{
Recuperar datos de un modelo logistf para emmeans
}
\keyword{internal}
