% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.modelo_germinacion_unificado}
\alias{summary.modelo_germinacion_unificado}
\title{Resumen de Modelo RegenerAR}
\usage{
\method{summary}{modelo_germinacion_unificado}(object, ...)
}
\arguments{
\item{object}{Objeto de clase modelo_germinacion_unificado.}

\item{...}{Argumentos adicionales.}
}
\value{
Objeto de clase summary.modelo_germinacion_unificado.
}
\description{
Genera un resumen estadístico para modelos de RegenerAR.
}
